move directory location of domain entity
