+++ src/main/java/com/spring/practice/rest/repository/impl/jpa/JpaUserEmRepository.java
@@ -7,7 +7,7 @@ import javax.persistence.EntityManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import com.spring.practice.rest.domain.User;
+import com.spring.practice.rest.domain.user.User;
 import com.spring.practice.rest.repository.UserRepository;
 
 @Repository("JpaUserEmRepository")
