+++ src/main/java/com/spring/practice/rest/repository/impl/jpa/JpaUserInterfaceRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.spring.practice.rest.domain.User;
+import com.spring.practice.rest.domain.user.User;
 
 @Repository("JpaUserInterfaceRepository")
 public interface JpaUserInterfaceRepository extends JpaRepository<User, Long> {
